requirements.txt
discord.py
python-dotenv
paramiko